fix extract logic bug and update

+++ api_server/routing/return_models/configures/node/overview.py
@@ -1,7 +1,7 @@
 from db.proxy import Node as Base
 from typing import Union, Optional
 extend_map = {
-    'position': Union[str, int]
+    'position': Union[str, float]
 
 }
 unpicks = [
